pyproject.toml
[tool.poetry]
name = "homework_drf"
version = "0.1.0"
description = ""
authors = ["KirillOreshchuk <oreshchuk.kirill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
faker = "^22.4.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
